Add files via upload
Resolved issue in sites table for groundwater change feature feature. The issue was identifying cells within sites table that have a value for groundwater change in the user specified interval. Added a test to only address the row in the sites table for sites that have a value for groundwater change in the user specified interval.